Converted to TypeScript
