+++ package.json
@@ -21,6 +21,7 @@
     "web-vitals": "^1.0.1"
   },
   "devDependencies": {
+    "@types/react-router-dom": "^5.1.7",
     "@typescript-eslint/eslint-plugin": "^4.22.0",
     "@typescript-eslint/parser": "^4.22.0",
     "eslint-config-airbnb": "^18.2.1",
